feature(routes/+page.svelte): almost rewrite the server part
