+++ .gitignore
@@ -15,6 +15,9 @@ Thumbs.db
 .env.*
 !.env.example
 !.env.test
+/wsServer/.env
+/wsServer/.env.*
+!/wsServer/.env.example
 
 # Vite
 vite.config.js.timestamp-*
