+++ src/components/GameRoomBar.svelte
@@ -1,5 +1,6 @@
 <script>
-	import RoomCode from '../components/joinRoom/RoomCode.svelte';
+	import RoomCode from './GameRoomBar/RoomCode.svelte';
+	import FailMessage from './GameRoomBar/FailMessage.svelte';
 	import Button from './Button.svelte';
 
 	export let status = {
@@ -7,6 +8,7 @@
 	};
 
 	export let room = {
+		failMessage: null,
 		yourRole: null
 	};
 
@@ -48,7 +50,10 @@
 		{#if status.game == 'waitForAnother'}
 			<span>房间号：{room.code}</span>
 		{/if}
-		<span class={room.yourRole == 'host' ? 'font-bold' : ''}>房主：{players.host}</span>
-		<span class={room.yourRole == 'client' ? 'font-bold' : ''}>参加者：{players.client}</span>
+		<span class={room.yourRole == 'host' ? 'underline' : ''}>房主：{players.host}</span>
+		<span class={room.yourRole == 'client' ? 'underline' : ''}>参加者：{players.client}</span>
+	{/if}
+	{#if room.failMessage != null}
+		<FailMessage message={room.failMessage} />
 	{/if}
 </fieldset>
