+++ wsServer/.env.example
@@ -1,0 +1,6 @@
+# 监听端口
+PORT=6789
+# 最大房间数
+MAX_ROOM=5
+# 生产环境
+PROD=false
