+++ wsServer/message.js
@@ -1,0 +1,9 @@
+export const sendMessage = (webSocket, type, content) => {
+	webSocket.send(
+		JSON.stringify({
+			v: '1',
+			type,
+			content
+		})
+	);
+};
