+++ wsServer/util.js
@@ -4,6 +4,21 @@ const roomNameTemplates = {
 	suffix: ['Java', 'CSharp', 'PHP', 'Go', 'Rust']
 };
 
+const nameTemplates = [
+	'Ryan Dahl',
+	'Brendan Eich',
+	'Michael Ficarra',
+	'Rich Harris',
+	'Isaac Z. Schlueter',
+	'Feross Aboukhadijeh',
+	'James M Snell',
+	'Wes Bos',
+	'Scott Tolinski',
+	'Shu-yu Guo',
+	'Jordan Harband',
+	'Matt Pocock'
+];
+
 let order = [];
 
 // 添加校验位
@@ -45,4 +60,22 @@ const generateRandomRoomNameAndCode = (index) => {
 	};
 };
 
-export { generateRandomRoomNameAndCode };
+// 生成0到value的随机整数（不包括value）
+const randomIntFromZeroTo = (value) => {
+	return Math.floor(Math.random() * value);
+};
+
+// 生成随机玩家名字
+const generateRandomPlayerName = (params) => {
+	let nameList = nameTemplates;
+	if (typeof params == 'object') {
+		const { except } = params;
+		if (except !== undefined) {
+			nameList = nameList.filter((one) => one != except);
+		}
+	}
+	const index = randomIntFromZeroTo(nameList.length);
+	return nameList[index];
+};
+
+export { generateRandomRoomNameAndCode, generateRandomPlayerName };
